requirements.txt
flask
supabase
python-dotenv
pandas
openpyxl
requests
